+++ lib/analog/src/lib.rs
@@ -1,5 +1,7 @@
 #![cfg_attr(not(test), no_std)]
 
+pub mod filter;
+
 use core::marker::Copy;
 use core::ops::{Add, Div, Mul};
 
@@ -27,43 +29,6 @@ where
     (count / max) * vref
 }
 
-#[derive(Debug, Clone)]
-pub struct MovingAvg<T, const N: usize>(heapless::HistoryBuf<T, N>);
-
-impl<const N: usize> MovingAvg<f32, N> {
-    /// Create a new moving average buffer.
-    pub fn new() -> Self {
-        Self(heapless::HistoryBuf::new())
-    }
-
-    /// Add a new value to the history buffer.
-    pub fn push(&mut self, value: f32) {
-        self.0.write(value);
-    }
-
-    /// Get the moving average.
-    ///
-    /// Only returns [`Some`] if one or more elements are in the buffer.
-    pub fn avg(&self) -> Option<f32> {
-        if self.0.is_empty() {
-            return None;
-        }
-
-        let mut value = 0.0;
-        for el in self.0.iter() {
-            value += el;
-        }
-
-        Some(value / self.0.len() as f32)
-    }
-}
-
-impl<const N: usize> Default for MovingAvg<f32, N> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -84,17 +49,5 @@ mod tests {
 
         let volts = count_to_volts(3.3, 1023.0, 0.0);
         assert_eq!(volts, 0.0);
-    }
-
-    #[test]
-    fn moving_average() {
-        let mut avg = MovingAvg::<f32, 32>::new();
-        assert!(avg.avg().is_none());
-
-        // comparing floats is not great, but good enough to smoke test this.
-        avg.push(0.0);
-        assert_eq!(avg.avg(), Some(0.0));
-        avg.push(1.0);
-        assert_eq!(avg.avg(), Some(0.5));
     }
 }